// light-bootstrap-dashboard-angular2-master/src/app/Model/SecPoint.ts
import { CommandePasserPar } from "./CommandePasserPar";

export class SecPoint {
    secPointId: number;
    client: string;
    dureeDeLicence: string;
    nomDuContact: string;
    adresseEmailContact: string;
    mailAdmin: string;
    ccMail: string[];
    commandePasserPar: CommandePasserPar;
    numero: string;
    approuve?: boolean;
    remarque: string;
    sousContrat: boolean;
    fichier?: string;
    fichierOriginalName?: string;
     licences: {
      nomDesLicences: string;
      quantite: string;
      dateEx: Date;
}[];
  
}